requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai>=1.0.0
tavily-python==0.2.6
requests==2.31.0
beautifulsoup4==4.12.2
rich==13.7.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
langchain==0.1.0
langgraph==0.0.10 
python-dotenv
requests